+++ server/services/index.ts
@@ -1,9 +1,11 @@
 import lifecycles from './lifecycles';
 import algolia from './algolia';
 import strapi from './strapi';
+import utils from './utils';
 
 export default {
   algolia,
   strapi,
   lifecycles,
+  utils,
 };