+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 
 
+## [1.6.0-0](https://github.com/wizbii/strapi-algolia/compare/v1.5.1...v1.6.0-0) (2024-04-22)
+
+
+### Features
+
+ - add hideFields property in configuration 
+ - add tests
+ - [fields] - hide fields configured
+ - [service] - add utils service
+
 ## [1.5.1](https://github.com/wizbii/strapi-algolia/compare/v1.5.0...v1.5.1) (2024-04-17)